added additional arguments for export_mx_to_file

+++ src/hdusd/utils/mx.py
@@ -249,12 +249,13 @@ def convert_mx_node_tree(context):
 
 
 def export_mx_to_file(doc, filepath, *, mx_node_tree=None, is_export_deps=False,
-                      is_export_textures=False, texture_dir_name="textures"):
+                      is_export_textures=False, texture_dir_name='textures',
+                      is_clean_texture_folder=True, is_clean_deps_folder=True):
     root_dir = Path(filepath).parent
 
     if is_export_deps and mx_node_tree:
         mx_libs_dir = root_dir / MX_LIBS_FOLDER
-        if os.path.isdir(mx_libs_dir):
+        if os.path.isdir(mx_libs_dir) and is_clean_deps_folder:
             shutil.rmtree(mx_libs_dir)
 
         # we need to export every deps only once
@@ -274,7 +275,7 @@ def export_mx_to_file(doc, filepath, *, mx_node_tree=None, is_export_deps=False,
 
     if is_export_textures:
         texture_dir = root_dir / texture_dir_name
-        if os.path.isdir(texture_dir):
+        if os.path.isdir(texture_dir) and is_clean_texture_folder:
             shutil.rmtree(texture_dir)
 
         Path(texture_dir).mkdir(parents=True, exist_ok=True)
